move Logging code to Logging directory
