+++ gemrb/core/Logging/Logger.cpp
@@ -16,9 +16,9 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-#include "System/Logger.h"
+#include "Logging/Logger.h"
 
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 
 #include <cstdio>
