+++ gemrb/core/Logging/Logger.h
