+++ gemrb/core/Logging/Loggers/Stdio.cpp
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-#include "System/Logger/Stdio.h"
+#include "Logging/Loggers/Stdio.h"
 #include "Streams/FileStream.h"
 
 #include <cstdio>
