+++ gemrb/core/Logging/Logging.h
@@ -28,7 +28,7 @@
 
 #include "exports.h"
 
-#include "System/Logger.h"
+#include "Logging/Logger.h"
 
 #include <cstdarg>
