+++ gemrb/plugins/GUIScript/PythonConversions.h
@@ -27,7 +27,7 @@
 #include "Sprite2D.h"
 #include "TableMgr.h"
 #include "SymbolMgr.h"
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 #include "Strings/String.h"
 
 namespace GemRB {
