+++ gemrb/plugins/MVEPlayer/gstmvedemux.h
@@ -24,7 +24,7 @@
 
 #include "globals.h"
 
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 
 using namespace GemRB;
