+++ gemrb/plugins/SDLVideo/SDLSurfaceSprite2D.cpp
@@ -24,7 +24,7 @@
 #include "SDLSurfaceDrawing.h"
 #include "SDLVideo.h"
 
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 
 namespace GemRB {
