+++ gemrb/plugins/TTFImporter/Freetype.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "Freetype.h"
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 
 namespace GemRB {
