+++ platforms/apple/AppleLogger.mm
@@ -26,7 +26,7 @@
 
 #include "AppleLogger.h"
 
-#include "System/Logging.h"
+#include "Logging/Logging.h"
 
 namespace GemRB {
 /*
