+++ platforms/vita/VitaLogger.cpp
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-#include "System/Logger/Vita.h"
+#include "VitaLogger.h"
 
 #include <psp2/kernel/clib.h>
