+++ platforms/windows/Win32Console.h
@@ -20,7 +20,7 @@
 #ifndef LOGGER_WIN32_H
 #define LOGGER_WIN32_H
 
-#include "System/Logger/Stdio.h"
+#include "Logging/Loggers/Stdio.h"
 
 #include <windows.h>
